Added contact history table and endpoint for all/single customer contact history. listener to new table and emit message been created

+++ server/customer.js
@@ -1,4 +1,5 @@
 var dl = require("./lib/dataLayer.js");
+var cfg = require("./cfg.json");
 
 var getCustomer = (req, res) => {
   if(req.params && req.params.id){
@@ -13,7 +14,7 @@ var getCustomer = (req, res) => {
     //no Id
     res.send(400, {"error": "BAD_REQUEST"});
   }
-}
+};
 
 var getAllCustomers = (req, res) => {
   dl.getAllCustomers((err, data) => {
@@ -22,8 +23,8 @@ var getAllCustomers = (req, res) => {
     } else {
       res.send(200, data);
     }
-  })
-}
+  });
+};
 
 var updateCustomer = (req, res) => {
   if(req.body){
@@ -38,10 +39,11 @@ var updateCustomer = (req, res) => {
     //no body
     res.send(400, {"error": "BAD_REQUEST"});
   }
-}
+};
 
 var createCustomer = (req, res) => {
-  if(req.body){
+  //creating a customer must include name of company
+  if(req.body && req.body.company){
     dl.createCustomer(req.body, (err) => {
       if(err) {
         res.send(err.code, err.obj);
@@ -53,22 +55,32 @@ var createCustomer = (req, res) => {
     //no body
     res.send(400, {"error": "BAD_REQUEST"});
   }
-}
+};
 
 var deleteCustomer = (req, res) => {
   if(req.params && req.params.id){
-    dl.deleteCustomer(req.params.id, (err) => {
-      if(err){
-        res.send(err.code, err.obj);
+    if(req.headers && req.headers.token){
+      if(req.headers.token === cfg.deletePassword){
+        dl.deleteCustomer(req.params.id, (err) => {
+          if(err){
+            res.send(err.code, err.obj);
+          } else {
+            res.send(202);
+          }
+        });
       } else {
-        res.send(202);
-      }
-    });
+        console.log(`Incorrect token '${req.headers.token}' when deleting customer ${req.params.id}`);
+        res.send(403, {"error": "INCORRECT_TOKEN"});
+      }      
+    } else {
+      console.log(`Missing Token when deleting customer ${req.params.id}`);
+      res.send(403, {"error": "INCORRECT_TOKEN"});
+    }
   } else {
     //no id
     res.send(400, {"error": "BAD_REQUEST"});
   }
-}
+};
 
 var customerContacted = (req, res) => {
   if(req.params && req.params.id){
@@ -80,17 +92,44 @@ var customerContacted = (req, res) => {
       }
     });
   } else {
-    //no body
+    //no parameter
     res.send(400, {"error": "BAD_REQUEST"});
   }
-}
+};
+
+var getAllContactHistory = (req, res) => {
+  dl.getAllContactHistory((err, result) => {
+    if(err) {
+      res.send(err.code, err.obj);
+    } else {
+      res.send(200, result);
+    }
+  });
+};
+
+var getCustomerContactHistory = (req, res) => {
+  if(req.params && req.params.id){
+    dl.getCustomerContactHistory(req.params.id, (err, result) => {
+      if(err) {
+        res.send(err.code, err.obj);
+      } else {
+        res.send(200, result);
+      }
+    });
+  } else {
+    //no parameter
+    res.send(400, {"error": "BAD_REQUEST"});
+  }
+};
 
 module.exports = (server) => {
-  server.get("/customer", getAllCustomers);
-  server.get("/customer/:id", getCustomer);
-  server.put("/customer", updateCustomer);
-  server.del("/customer/:id", deleteCustomer);
-  server.post("/customer", createCustomer);
+  server.get("/customer/contact", getAllContactHistory);
   server.put("/customer/contact/:id", customerContacted);
+  server.get("/customer/contact/:id", getCustomerContactHistory);
+  server.get("/customer", getAllCustomers);
+  server.put("/customer", updateCustomer);
+  server.post("/customer", createCustomer);
+  server.get("/customer/:id", getCustomer);
+  server.del("/customer/:id", deleteCustomer);
   require("./lib/dataEventListener.js")(server, dl);
 }
